requirements.txt
# DAT Service Library — Python 3.8 compatible
# Install: pip install -r requirements.txt

# Core (no external deps for domain layer — by design)
# The domain layer uses only stdlib + dataclasses

# Adapters (install what you need)
psycopg2-binary>=2.9.0     # PostgreSQL adapter
pydantic>=1.10.0,<2.0.0    # Validation (1.x for 3.8 compat)

# Observability
structlog>=21.0.0           # Structured logging
prometheus-client>=0.14.0   # Metrics

# Auth
pyjwt>=2.4.0                # JWT validation

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Data analysis (optional — for sensor analytics)
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
